move quiz creation side panels to own subdir, fix refs
