+++ kolibri/plugins/coach/assets/src/views/quizzes/CreateExamPage/SidePanels/ReplaceQuestions.vue
@@ -157,8 +157,8 @@
   import AccordionContainer from 'kolibri-common/components/AccordionContainer';
   import useAccordion from 'kolibri-common/components/useAccordion';
   import useSnackbar from 'kolibri/composables/useSnackbar';
-  import { injectQuizCreation } from '../../../composables/useQuizCreation';
-  import { PageNames } from '../../../constants/index';
+  import { injectQuizCreation } from '../../../../composables/useQuizCreation';
+  import { PageNames } from '../../../../constants/index';
 
   export default {
     name: 'ReplaceQuestions',
