+++ kolibri/plugins/coach/assets/src/views/quizzes/CreateExamPage/SidePanels/SectionEditor.vue
@@ -134,9 +134,9 @@
   import { MAX_QUESTIONS_PER_QUIZ_SECTION } from 'kolibri/constants';
   import useKResponsiveWindow from 'kolibri-design-system/lib/composables/useKResponsiveWindow';
   import useSnackbar from 'kolibri/composables/useSnackbar';
-  import { PageNames } from '../../../constants/index';
-  import { injectQuizCreation } from '../../../composables/useQuizCreation.js';
-  import useDrag from './useDrag.js';
+  import { PageNames } from '../../../../constants/index';
+  import { injectQuizCreation } from '../../../../composables/useQuizCreation.js';
+  import useDrag from '../useDrag.js';
 
   export default {
     name: 'SectionEditor',
